requirements.txt
Django>=3.1.3,<3.2.0
Pillow>=8.0.1,<8.2.0
djangorestframework>=3.12.2,<3.13.0
markdown>=3.3.3,<3.4.0
django-filter>=2.4.0,<2.5.0
djangorestframework-simplejwt>=4.6.0,<4.7.0
